mev_bot/requirements.txt
web3
requests
psycopg2-binary
python-telegram-bot
discord.py
pandas